修改md5.h头文件解决64位MD5计算不对问题；更新url库去掉错误的url

+++ md5/md5.h
@@ -26,10 +26,10 @@ documentation and/or software.
 typedef unsigned char *POINTER;
 
 /* UINT2 defines a two byte word */
-typedef unsigned short int UINT2;
+typedef unsigned short UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef unsigned int UINT4;
 
 /* MD5 context. */
 typedef struct {
